# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.ConfirmPaymentRequest:
    properties:
      transaction_id:
        example: txn-987654
        type: string
    type: object
...
  handler.CreatePaymentRequest:
    properties:
      amount:
        example: 1500.5
        type: number
      order_id:
        example: ord-123456
        type: string
    type: object
  handler.CreateProductRequest:
    properties:
      category_ids:
        example:
        - cat-1
        - cat-2
        items:
          type: string
        type: array
      description:
        example: High precision RGB gaming mouse
        type: string
      image_urls:
        example:
        - http://example.com/img1.png
        - http://example.com/img2.png
        items:
          type: string
        type: array
      name:
        example: Gaming Mouse
        type: string
      price:
        example: 1250
        type: number
      status:
        example: active
        type: string
      stock:
        example: 50
        type: integer
    type: object
...
  handler.UpdateProductRequest:
    properties:
      description:
        example: Upgraded version
        type: string
      name:
        example: Gaming Mouse V2
        type: string
      price:
        example: 1500
        type: number
      status:
        example: active
        type: string
      stock:
        example: 30
        type: integer
    type: object
...
  phajay.WebhookPayload:
    properties:
      amount:
        type: number
      orderNo:
        type: string
      status:
        description: success, failed, cancelled
        type: string
      transactionId:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample e-commerce server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Golang E-Commerce API
  version: "1.0"
paths:
...
  /admin/payments/{id}/confirm:
    patch:
      consumes:
      - application/json
      description: Allows admin to manually confirm a payment
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      - description: Transaction details (transaction_id)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ConfirmPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Confirm payment manually (Admin)
      tags:
      - admin
      - payments
...
  /products:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all products
      tags:
      - products
  /products/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve details of a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get product by ID
      tags:
      - products
  /products/category/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve all products in a specific category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get products by category
      tags:
      - products
  /products/seller/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve all products listed by a specific seller
      parameters:
      - description: Seller ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get products by seller
      tags:
      - products
...
  /seller/products/create:
    post:
      consumes:
      - application/json
      description: Create a new product listed by the current seller
      parameters:
      - description: Product details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new product (Seller)
      tags:
      - products
      - seller
  /seller/products/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product listed by the current seller
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a product (Seller)
      tags:
      - products
      - seller
  /seller/products/update/{id}:
    put:
      consumes:
      - application/json
      description: Update details of a product listed by the current seller
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated product details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a product (Seller)
      tags:
      - products
      - seller
...
  /user/payments:
    post:
      consumes:
      - application/json
      description: Create a Phajay payment link for an order
      parameters:
      - description: Payment details (order_id, amount)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a payment
      tags:
      - payments
  /user/payments/order/{orderId}:
    get:
      consumes:
      - application/json
      description: Get payment info for a specific order
      parameters:
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get payment by order
      tags:
      - payments
...
  /webhooks/phajay:
    post:
      consumes:
      - application/json
      description: Handle the Phajay webhook callback (Public)
      parameters:
      - description: Webhook payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/phajay.WebhookPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Phajay payment webhook
      tags:
      - payments
      - webhooks
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"